Map, player fini

+++ MazeRunner-Netbeans/src/Graphics/View.java
@@ -19,7 +19,7 @@ import static org.lwjgl.opengl.GL11.glVertex2f;
 public class View {
    
     
-    public static void quadData(int x,int y,int w,int h,float[] color,int xo, int yo){
+    public static void quadData(float x,float y,int w,int h,float[] color,int xo, int yo){
         glColor4f(color[0],color[1],color[2],color[3]);
         glTexCoord2f((0 + xo)/32.0f,(0 + yo)/32.0f); glVertex2f(x,y);
         glTexCoord2f((1 + xo)/32.0f,(0 + yo)/32.0f); glVertex2f(x + w,y);
@@ -27,13 +27,13 @@ public class View {
         glTexCoord2f((0 + xo)/32.0f,(1 + yo)/32.0f); glVertex2f(x,y+h);
     }
     
-    public static void renderQuad(int x,int y,int w,int h, float[] color, int xo, int yo){
+    public static void renderQuad(float x,float y,int w,int h, float[] color, int xo, int yo){
         glBegin(GL_QUADS);
         quadData(x,y,w,h,color, xo , yo);
         glEnd();
     }
     
-    public static void renderCharacters (int x,int y,int w,int h, float[] color,float size, int xo, int yo){
+    public static void renderCharacters (float x,float y,int w,int h, float[] color,float size, int xo, int yo){
         glBegin(GL_QUADS);
         glColor4f(color[0],color[1],color[2],color[3]);
         glTexCoord2f((0 + xo)/size,(0 + yo)/size); glVertex2f(x,y);
